+++ tests/testthat/test-class_options.R
@@ -398,9 +398,13 @@ tar_test("controller", {
   skip_if_not_installed("crew")
   x <- options_init()
   expect_silent(x$validate_controller(NULL))
-  expect_silent(x$validate_controller(crew::crew_controller_local()))
+  expect_silent(
+    x$validate_controller(
+      crew::crew_controller_local(host = "127.0.0.1")
+    )
+  )
   expect_null(x$get_controller())
-  x$set_controller(crew::crew_controller_local())
+  x$set_controller(crew::crew_controller_local(host = "127.0.0.1"))
   expect_true(inherits(x$get_controller(), "crew_class_controller"))
   x$reset()
   expect_null(x$get_controller())