+++ tests/testthat/test-tar_make.R
@@ -27,7 +27,10 @@ tar_test("tar_make() works with crew", {
   }
   tar_script({
     tar_option_set(
-      controller = crew::crew_controller_local(seconds_interval = 0.5),
+      controller = crew::crew_controller_local(
+        host = "127.0.0.1",
+        seconds_interval = 0.5
+      ),
       backoff = tar_backoff(min = 0.5, max = 0.5)
     )
     tar_target(