+++ tests/testthat/test-tar_option_set.R
@@ -313,7 +313,9 @@ tar_test("seed", {
 tar_test("controller", {
   skip_if_not_installed("crew")
   expect_null(tar_option_get("controller"))
-  tar_option_set(controller = crew::crew_controller_local())
+  tar_option_set(
+    controller = crew::crew_controller_local(host = "127.0.0.1")
+  )
   expect_true(inherits(tar_option_get("controller"), "crew_class_controller"))
   tar_option_reset()
   expect_null(tar_option_get("controller"))